@@ -1,7 +1,7 @@
 package me.barta.stayintouch.common.ui
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 
 /**
  * Base Activity implementing MVP pattern
