@@ -2,7 +2,7 @@ package me.barta.stayintouch.common.ui
 
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.View
 
 /**
@@ -17,7 +17,7 @@ abstract class MVPFragment<V : MVPContract.View, out P : MVPContract.Presenter<V
     protected abstract fun createComponent(): C
 
     @Suppress("UNCHECKED_CAST")
-    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         presenter.attachView(this as V)
     }
