@@ -1,9 +1,9 @@
 package me.barta.stayintouch.contactlist
 
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentPagerAdapter
-import android.support.v4.app.FragmentStatePagerAdapter
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentPagerAdapter
+import androidx.viewpager2.adapter.FragmentStateAdapter
 
 import me.barta.stayintouch.datastore.models.ContactCategory
 
@@ -11,12 +11,9 @@ import me.barta.stayintouch.datastore.models.ContactCategory
  * A [FragmentPagerAdapter] that returns a fragment corresponding to
  * one of the sections/tabs/pages.
  */
-class SectionsPagerAdapter(fm: FragmentManager, private val contactCategories: List<ContactCategory>,
-                           private val FragmentProviderFor: (Int) -> Fragment) : FragmentStatePagerAdapter(fm) {
+class SectionsPagerAdapter(activity: AppCompatActivity, private val contactCategories: List<ContactCategory>,
+                           private val FragmentProviderFor: (Int) -> Fragment) : FragmentStateAdapter(activity) {
 
-    override fun getItem(position: Int): Fragment = FragmentProviderFor(position)
-
-    override fun getCount(): Int = contactCategories.size
-
-    override fun getPageTitle(position: Int): CharSequence = contactCategories[position].name
+    override fun getItemCount(): Int = contactCategories.size
+    override fun createFragment(position: Int): Fragment = FragmentProviderFor(position)
 }
